Moved DoubleLinkedIntListItem to commons project (where it is used by the tests).
