+++ jsettlers.common/src/test/java/jsettlers/common/buildings/BuildingConfigurationsTest.java
@@ -13,7 +13,6 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
-import jsettlers.testutils.TestUtils;
 import jsettlers.common.buildings.stacks.RelativeStack;
 import jsettlers.common.position.RelativePoint;
