+++ jsettlers.common/src/test/java/jsettlers/common/utils/collections/list/DoubleLinkedListTest.java
@@ -22,7 +22,6 @@ import java.io.IOException;
 import java.util.Iterator;
 
 import jsettlers.testutils.TestUtils;
-import jsettlers.algorithms.path.astar.queues.bucket.DoubleLinkedIntListItem;
 
 import org.junit.Test;
